Moved `HTTPClientSpy` to a separate `Helpers` group (inside `Feed API` test folder)
