+++ EssentialFeedTests/Feed API/Helpers/HTTPClientSpy.swift
@@ -1,0 +1,25 @@
+import Foundation
+import EssentialFeed
+
+class HTTPClientSpy: HTTPClient {
+    var messages = [(url: URL, completion: (HTTPClientResult) -> Void)]()
+    var requestedURLs: [URL] { return messages.map { $0.url } }
+
+    func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void) {
+        messages.append((url, completion))
+    }
+
+    func completeWith(error: Error, at index: Int = 0) {
+        messages[index].completion(.failure(error))
+    }
+
+    func completeWith(statusCode: Int, data: Data, at index: Int = 0) {
+        let response = HTTPURLResponse(
+            url: messages[index].url,
+            statusCode: statusCode,
+            httpVersion: nil,
+            headerFields: nil
+        )!
+        messages[index].completion(.success((data, response)))
+    }
+}
